L1-support-tools/05-it-knowledge-cli/windows/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "it-knowledge-cli"
version = "1.0.0"
description = "Offline-capable IT knowledge base CLI for L1 support (Windows)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
dependencies = [
    "click>=8.0",
    "python-frontmatter>=1.0",
    "mistune>=3.0",
    "requests>=2.28",
    "numpy>=1.24",
    "sqlite-vec>=0.1.0",
]

[project.scripts]
it-kb = "it_kb.__main__:cli"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pyinstaller>=5.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["it_kb*"]
